@@ -1,14 +1,11 @@
-﻿namespace EF003.ExecuteInsertReturnIdentity
+﻿namespace EF003.ExecuteInsertReturnIdentity;
+public class Wallet
 {
-    public class Wallet
+    public int Id { get; set; }
+    public string? Holder { get; set; }
+    public decimal? Balance { get; set; }
+    public override string ToString()
     {
-        public int Id { get; set; }
-        public string? Holder { get; set; }
-        public decimal? Balance { get; set; }
-
-        public override string ToString()
-        {
-            return $"[{Id}] {Holder} ({Balance:C})";
-        }
+        return $"[{Id}] {Holder} ({Balance:C})";
     }
 }
